antifraud_af/requirements.txt
fastapi
uvicorn
httpx
redis
python-dotenv
requests